Add confirmation before delete subject

+++ SubjectModuleUI/DeleteSubject/DeleteSubjectForm.cs
@@ -34,17 +34,26 @@ namespace SubjectModuleUI.DeleteSubject
         {
             try
             {
-                var selectedSubject = this.comboBoxSelectSubjectToDelete.SelectedItem as Subject;
-                ClassFactory.GetOrCreate<SubjectLogic>().DeleteSubjectByCode(selectedSubject.Code);
-                this.labelActionResult.Text = "Subject " + selectedSubject.Name + " was succesfully deleted";
-                this.labelActionResult.Visible = true;
+                if(UserConfirmsThatWantToDeleteSubject())
+                {
+                    var selectedSubject = this.comboBoxSelectSubjectToDelete.SelectedItem as Subject;
+                    ClassFactory.GetOrCreate<SubjectLogic>().DeleteSubjectByCode(selectedSubject.Code);
+                    this.labelActionResult.Text = "Subject " + selectedSubject.Name + " was succesfully deleted";
+                    this.labelActionResult.Visible = true;
+                }
             }
             catch(Exception ex)
             {
                 this.labelActionResult.Text = ex.Message;
                 this.labelActionResult.Visible = true;
             }
-            
+        }
+
+        private bool UserConfirmsThatWantToDeleteSubject()
+        {
+            var userAnswer = MessageBox.Show("Are you sure you want to delete this subject?",
+                    "Delete Subject", MessageBoxButtons.YesNo, MessageBoxIcon.Information);
+            return userAnswer == DialogResult.Yes;
         }
 
         private void buttonCancel_Click(object sender, EventArgs e)
